Moved headers to a seperate folder
Contributes to CURA-9365
