+++ include/Savitar/MetadataEntry.h
