+++ src/Face.cpp
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Face.h"
+#include "Savitar/Face.h"
 
 using namespace Savitar;
