+++ src/MeshData.cpp
@@ -16,8 +16,8 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "MeshData.h"
-#include "../pugixml/src/pugixml.hpp"
+#include "Savitar/MeshData.h"
+#include <pugixml.hpp>
 #include <cstring>
 #include <iostream>
 #include <stdexcept> //For std::runtime_error.
