+++ src/Vertex.cpp
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Vertex.h"
+#include "Savitar/Vertex.h"
 
 using namespace Savitar;
